first commitftdfjfff

+++ solution.py
@@ -170,7 +170,7 @@ def get_route(hostname):
                     timeSent = struct.unpack("d", recvPacket[28:28 + bytes])[0]
                     #Fill in start
                     tracelist1.append(ttl)
-                    tracelist1.append(str(int((timeReceived - t)*1000))+"ms")
+                    tracelist1.append(str((timeReceived - t)*1000)+"ms")
                     tracelist1.append(str(addr[0]))
                     tracelist1.append(resolved)
                     tracelist2.append(tracelist1)
@@ -181,7 +181,7 @@ def get_route(hostname):
                     timeSent = struct.unpack("d", recvPacket[28:28 + bytes])[0]
                     #Fill in start
                     tracelist1.append(ttl)
-                    tracelist1.append(str(int((timeReceived - t)*1000))+"ms")
+                    tracelist1.append(str((timeReceived - t)*1000)+"ms")
                     tracelist1.append(str(addr[0]))
                     tracelist1.append(str(resolved))
                     tracelist2.append(tracelist1)
@@ -192,7 +192,7 @@ def get_route(hostname):
                     timeSent = struct.unpack("d", recvPacket[28:28 + bytes])[0]
                     #Fill in start
                     tracelist1.append(ttl)
-                    tracelist1.append(str(int((timeReceived - t)*1000))+"ms")
+                    tracelist1.append(str((timeReceived - t)*1000)+"ms")
                     tracelist1.append(str(addr[0]))
                     tracelist1.append(str(resolved))
                     tracelist2.append(tracelist1)
@@ -216,6 +216,4 @@ def get_route(hostname):
         #print(tracelist2)
     return tracelist2
     
-
-
 get_route("www.cnn.com")
